requirements.txt
ansible==1.9.4
asn1crypto==0.23.0
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.4
cx-Oracle==6.0.3
Django==1.11.7
djangorestframework==3.7.3
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
Jinja2==2.10
MarkupSafe==1.0
MySQL-python==1.2.5
paramiko==2.4.0
passlib==1.7.1
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.2.0
pytz==2017.3
PyYAML==3.12
six==1.11.0
uWSGI==2.0.15
xlrd==1.1.0
